==> 2.ipynb <==
import pandas as pd
import numpy as np

from utils import read_csv
from config import DATA_PATH
# from pipeline import dblp_paper_df, asn_paper_df, author_df, coauthor_df

# It will be faster if we load processed data, for repeat test, its recommended to use loaded data instead of pipeline.

dblp_paper_df = read_csv(DATA_PATH, 'dblp_paper.csv')
# asn_paper_df = read_csv(DATA_PATH, 'asn_paper.csv', index_col='id')
# author_df = read_csv(DATA_PATH, 'author.csv', index_col='author_id')
# coauthor_df = read_csv(DATA_PATH, 'coauthor.csv')

dblp_paper_df1 = dblp_paper_df.dropna(subset=["authors_org"])
dblp_paper_df1['affiliation'] = dblp_paper_df1['authors_org'].apply(lambda s: s.split(';')[0] if s else None)

from temp import aff2country
def overview_global(paper_df):
    df = paper_df[['id', 'year', 'affiliation']].copy()
    df['country'] = df['affiliation'].apply(aff2country)
    df = df.dropna(subset=['country'])
    grouped_paper = df.groupby(["year", 'country'])
    df1 = pd.DataFrame()
    count = grouped_paper["id"].apply(len)
    df1["year"] = [idx[0] for idx in count.axes[0]]
    df1['country'] = [idx[1] for idx in count.axes[0]]
    df1["publication count"] = count.values
    return df1

df1 = overview_global(dblp_paper_df1)

def keywords_freq(paper_df, year):
    """ Calculate keywords frequency according to paper_df

    For now, only dblp dataset contains keywords column
    Args:
        paper_df: pd.Dataframe
            dblp data
        year: int
            year to calculate keyword frequency
    Returns:
        pd.Dataframe
        year | keyword1 | keyword2 | ...
        ________________________________
             |          |          | ...
    """
    keywords = paper_df[paper_df['year'] == year]['keywords'].values
    count = dict()
    for l in keywords:
        for k in l.split(';'):
            count[k] = count.get(k, 0) + 1
    return count

keywords_freq(dblp_paper_df, 2010)
