==> 4.ipynb <==
from process import *

coauthor_df = read_csv(DATA_PATH, 'coauthor.csv')
author_df = read_csv(DATA_PATH, 'processed_author.csv')

city_cooperate = cooperate_strength(coauthor_df, author_df, by='affiliation')

city_cooperate.to_csv(os.path.join(DATA_PATH, 'affiliation_cooperate.csv'), index=False)

affs = pd.unique(author_df['affiliation'])

df = pd.DataFrame()
df['aff'] = affs


df.to_csv('affs.csv', index=False, encoding='utf-8')

df.to_excel()

aff_ = [str(i).split(',')[0] for i in affs]

aff_1 = pd.unique(aff_)

paper_df = read_csv(DATA_PATH, 'dblp_paper.csv')

paper_df['fos_name'].isna().sum()

authors = read_csv(DATA_PATH, 'processed_author.csv')

sorted_authors = authors.sort_values('H-index')
