==> 6.ipynb <==
from process import *
import re
from collections import Counter
from utils import *

paper_df = read_csv(DATA_PATH, 'processed_dblp_paper.csv')

df = pd.DataFrame()
df['aff'] = paper_df['affiliation'].astype('str').apply(lambda s: re.sub('#.*?#', '', s))

df['aff'] = df['aff'].apply(lambda x: x.lower())
df['aff'] = multiprocess_str(split_data(df, size=2000), 'aff')

def process(line):
    strings = line.split(',')
    def extract(string):
        pat = re.compile('[a-z]+')
        string_ = ' '.join(re.findall(pat, string))
        return string_
    return ','.join([extract(string) for string in strings if extract(string)])
        
df['processed_aff'] = df['aff'].apply(process)

pd.unique(df['processed_aff'].apply(lambda x: re.findall('univ[a-z]+', x)[0] if re.findall('univ[a-z]+', x) else np.nan))

reg_ = pd.unique(df['processed_aff'].apply(lambda x: x.split(',')[-1]))
reg__ = np.array([x for x in reg_ if len(x.split())==1])

def split_univ(aff):
    for x in aff.split(','):
        for x_ in x.split():
            if x_.startswith('univ'):
                return x

df['univ1'] = df['processed_aff'].apply(split_univ)

df['univ1'].isna().sum()

univs = pd.unique(df['univ1'])

univ_df = pd.DataFrame()
univ_df['university'] = univs

chrs = set()
chrs = chrs.union(*df['aff'].apply(set))

for i in list(chrs)[680:]:
    print(i)

word_count = Counter(' '.join(df['processed_aff'].values).replace(",",' ').split(' '))

univ_count = df['univ1'].value_counts()

univ_count.to_csv('univ_count.csv')

univ_count[:700].sum()

word_count_df = pd.DataFrame()
word_count_df['word'] = word_count.keys()
word_count_df['count'] = word_count.values()

word_count_df = word_count_df.sort_values('count', ascending=False).reset_index(drop=True)

paper
