==> 7.ipynb <==
from process import *
import re
from collections import Counter
from utils import *

def h_index(df):
    processed_df = df.sort_values(by=['n_cites'], ascending=False).copy().reset_index(drop=True)
    for i in range(0, processed_df.shape[0]):
        if int(processed_df['n_cites'][i]) <= i:
            return i
    return processed_df.shape[0]

paper_df_ = paper_df.dropna(subset=['country'])
grouped_paper = paper_df_.groupby(by='country')

country_h = grouped_paper.apply(h_index)

region_str(paper_df)

for name, group in grouped_paper:
    print()



l = {name: country_H(paper_df,name) for name in country_name if name}

c = pd.unique(paper_df['country'].astype(str))
